// src/types/enums.ts
export enum FILTERS {
  all = 'All',
  completed = 'Completed',
  active = 'Active',
};

export enum  ERRORS_MESSAGE {
  loadError = 'Unable to load todos',
  titleError = 'Title should not be empty',
  addingError = 'Unable to add a todo',
  deleteError = 'Unable to delete a todo',
  updateError = 'Unable to update a todo',
};